Refactor Drawer
